redesigned template for SpaceTraders
