+++ SpaceTraders/.gitignore
