+++ SpaceTraders/app/.gitignore
