+++ SpaceTraders/app/src/main/res/mipmap-mdpi/ic_launcher.png
