+++ SpaceTraders/gradle.properties
