+++ SpaceTraders/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Feb 12 00:56:39 EST 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.1-all.zip
